Edits
Where we tell users to click their initials, I added "SN in the figure below" to provide a visual cue to users. In the table, for field 2, in the sentence "All properties that do not contain..." I replaced "properties" with "users or roles." In the bottom table, I provided a link for creating new users to compliment the link that was already defined for creating new roles.